requirements.txt
streamlit
pydantic
google-generativeai
duckduckgo-search
beautifulsoup4
requests
python-dotenv
pyyaml
watchdog
